--HG--
branch : emailMessages

+++ app/protected/modules/zurmo/components/WebUser.php
@@ -55,15 +55,17 @@
                     return parent::__get($attributeName);
                 }
             }
-            catch(NotFoundException $e)
+            catch (NotFoundException $e)
             {
-                //Perhaps the username has changed, clear session and logout user.
-                Yii::app()->getSession()->destroy();
-                $this->redirect(Yii::app()->homeUrl);
+                if (Yii::app()->isApplicationInstalled())
+                {
+                    //Perhaps the username has changed, clear session and logout user.
+                    Yii::app()->getSession()->destroy();
+                    Yii::app()->request->redirect(Yii::app()->homeUrl);
+                }
             }
             catch (CException $e)
             {
-
             }
         }
